Tighten Learn Zen guide and docs truth tests
