+++ tests/docs_truth/public_docs/diagnostics_catalog.rs
@@ -1,0 +1,124 @@
+use super::super::*;
+
+#[test]
+fn diagnostics_catalog_documents_json_stable_codes() {
+    let catalog = read("docs/DIAGNOSTICS.md");
+
+    for required in [
+        "# Zen Diagnostics Catalog",
+        "JSON-Stable Codes",
+        "E2000",
+        "removed syntax or reserved syntax",
+        "replace_removed_return_with_final_expression",
+        "feature_gate",
+        "E3500",
+        "Sync/Async effect modes",
+        "allocator-backed dynamic `String`",
+        "std actor framework types",
+        "std actor framework types/imports",
+        "std allocator imports",
+        "std Sync/Async runtime imports",
+        "E0203",
+        "gated compiler-owned intrinsic call",
+        "comptime type matching",
+        "reserved async scheduler intrinsics",
+        "atomic intrinsics",
+        "raw syscalls",
+        "raw allocation intrinsics",
+        "byte-memory intrinsics",
+        "raw pointer intrinsics",
+        "E3053",
+        "gated range expression",
+        "E3054",
+        "gated Result propagation",
+        "E3055",
+        "gated task waiting",
+        "resolver validation failure",
+        "gated reserved type surfaces",
+        "E5000",
+        "generic inference conflict",
+        "E5001",
+        "generic type-argument arity",
+        "behavior references",
+        "E5002",
+        "type arguments were supplied to a non-generic",
+        "E6004",
+        "generic behavior-bound failure",
+        "E6007",
+        "explicit type association `.requires` failure",
+        "E6010",
+        "behavior implementation coherence failure",
+        "tests/fixtures/ir_json/diagnostics_return.golden.json",
+        "tests/fixtures/ir_json/diagnostics_type_match_gate.golden.json",
+        "tests/fixtures/ir_json/diagnostics_async_intrinsic_gate.golden.json",
+        "tests/fixtures/ir_json/diagnostics_atomic_gate.golden.json",
+        "tests/fixtures/ir_json/diagnostics_syscall_gate.golden.json",
+        "tests/fixtures/ir_json/diagnostics_raw_allocate_gate.golden.json",
+        "tests/fixtures/ir_json/diagnostics_byte_memory_gate.golden.json",
+        "tests/fixtures/ir_json/diagnostics_raw_pointer_gate.golden.json",
+        "tests/fixtures/ir_json/diagnostics_range_gate.golden.json",
+        "tests/fixtures/ir_json/diagnostics_raise_gate.golden.json",
+        "tests/fixtures/ir_json/diagnostics_await_gate.golden.json",
+        "tests/fixtures/ir_json/diagnostics_duplicate_generic_requires.golden.json",
+        "tests/fixtures/ir_json/diagnostics_duplicate_generic_impl.golden.json",
+        "tests/fixtures/ir_json/diagnostics_dynamic_string_gate.golden.json",
+        "tests/fixtures/ir_json/diagnostics_generic_dynamic_string_gate.golden.json",
+        "tests/fixtures/ir_json/diagnostics_typed_allocator_effect_gate.golden.json",
+        "tests/fixtures/ir_json/diagnostics_sync_effect_gate.golden.json",
+        "tests/fixtures/ir_json/diagnostics_async_effect_gate.golden.json",
+        "tests/fixtures/ir_json/diagnostics_actor_type_gate.golden.json",
+        "tests/fixtures/ir_json/diagnostics_bare_actor_type_gate.golden.json",
+        "tests/fixtures/ir_json/diagnostics_actor_import_gate.golden.json",
+        "tests/fixtures/ir_json/diagnostics_allocator_import_gate.golden.json",
+        "tests/fixtures/ir_json/diagnostics_async_runtime_import_gate.golden.json",
+        "tests/fixtures/ir_json/diagnostics_sync_runtime_import_gate.golden.json",
+        "tests/fixtures/ir_json/diagnostics_generic_function_arity.golden.json",
+        "tests/fixtures/ir_json/diagnostics_generic_function_type_arg_annotation_arity.golden.json",
+        "tests/fixtures/ir_json/diagnostics_generic_method_type_arg_annotation_arity.golden.json",
+        "tests/fixtures/ir_json/diagnostics_generic_method_type_arg_annotation_missing_args.golden.json",
+        "tests/fixtures/ir_json/diagnostics_closure_param_annotation_type_arg_arity.golden.json",
+        "tests/fixtures/ir_json/diagnostics_closure_return_annotation_missing_args.golden.json",
+        "tests/fixtures/ir_json/diagnostics_cast_target_annotation_type_arg_arity.golden.json",
+        "tests/fixtures/ir_json/diagnostics_cast_target_annotation_missing_args.golden.json",
+        "tests/fixtures/ir_json/diagnostics_nested_generic_annotation_inner_arity.golden.json",
+        "tests/fixtures/ir_json/diagnostics_nested_generic_instantiation_inner_arity.golden.json",
+        "tests/fixtures/ir_json/diagnostics_function_type_parameter_annotation_arity.golden.json",
+        "tests/fixtures/ir_json/diagnostics_function_type_return_annotation_missing_args.golden.json",
+        "tests/fixtures/ir_json/diagnostics_pointer_inner_generic_annotation_arity.golden.json",
+        "tests/fixtures/ir_json/diagnostics_slice_inner_generic_annotation_missing_args.golden.json",
+        "tests/fixtures/ir_json/diagnostics_array_inner_generic_annotation_arity.golden.json",
+        "tests/fixtures/ir_json/diagnostics_generic_struct_local_annotation_arity.golden.json",
+        "tests/fixtures/ir_json/diagnostics_generic_struct_local_annotation_missing_args.golden.json",
+        "tests/fixtures/ir_json/diagnostics_generic_enum_local_annotation_arity.golden.json",
+        "tests/fixtures/ir_json/diagnostics_generic_enum_local_annotation_missing_args.golden.json",
+        "tests/fixtures/ir_json/diagnostics_generic_struct_constructor_arity.golden.json",
+        "tests/fixtures/ir_json/diagnostics_generic_struct_constructor_missing_args.golden.json",
+        "tests/fixtures/ir_json/diagnostics_generic_enum_constructor_arity.golden.json",
+        "tests/fixtures/ir_json/diagnostics_generic_enum_constructor_missing_args.golden.json",
+        "tests/fixtures/ir_json/diagnostics_generic_struct_annotation_arity.golden.json",
+        "tests/fixtures/ir_json/diagnostics_generic_enum_annotation_arity.golden.json",
+        "tests/fixtures/ir_json/diagnostics_generic_struct_annotation_missing_args.golden.json",
+        "tests/fixtures/ir_json/diagnostics_generic_enum_annotation_missing_args.golden.json",
+        "tests/fixtures/ir_json/diagnostics_generic_result_method_arity.golden.json",
+        "tests/fixtures/ir_json/diagnostics_generic_requires_arity.golden.json",
+        "tests/fixtures/ir_json/diagnostics_generic_impl_arity.golden.json",
+        "tests/fixtures/ir_json/diagnostics_generic_extends_arity.golden.json",
+        "tests/fixtures/ir_json/diagnostics_nongeneric_struct_annotation_type_args.golden.json",
+        "tests/fixtures/ir_json/diagnostics_nongeneric_enum_annotation_type_args.golden.json",
+        "tests/fixtures/ir_json/diagnostics_nongeneric_struct_constructor_type_args.golden.json",
+        "tests/fixtures/ir_json/diagnostics_nongeneric_enum_constructor_type_args.golden.json",
+        "tests/fixtures/ir_json/diagnostics_nongeneric_function_type_args.golden.json",
+        "tests/fixtures/ir_json/diagnostics_nongeneric_module_function_type_args.golden.json",
+        "tests/fixtures/ir_json/diagnostics_nongeneric_builtin_function_type_args.golden.json",
+        "tests/fixtures/ir_json/diagnostics_nongeneric_method_type_args.golden.json",
+        "tests/fixtures/ir_json/diagnostics_generic_result_method_bound.golden.json",
+        "tests/fixtures/ir_json/diagnostics_generic_requires_missing_impl.golden.json",
+        "tests/fixtures/ir_json/diagnostics_generic_result_method_inference.golden.json",
+        "tests/fixtures/ir_json/diagnostics_generic_behavior_overlap.golden.json",
+    ] {
+        assert!(
+            catalog.contains(required),
+            "docs/DIAGNOSTICS.md is missing diagnostic catalog text: {required}"
+        );
+    }
+}
