+++ tests/docs_truth/public_docs/learn_zen.rs
@@ -1,0 +1,212 @@
+use super::super::*;
+
+#[test]
+fn learn_zen_guide_covers_core_tour_and_gated_previews() {
+    let guide = read("docs/learn_zen_in_y_minutes.md");
+
+    for required in [
+        "## Comments",
+        "prefix-first declarations",
+        "stable source forms you can use in examples today",
+        "gated design previews that show intended syntax",
+        "Stable Zen is deliberately small",
+        "Need | Spell it like this",
+        "Copy These Forms First",
+        "## Control Flow At A Glance",
+        "No alternate loop syntax",
+        "Keyword-style forms from other languages",
+        "Translation Cheat Sheet",
+        "`while condition { ... }`",
+        "`for item in items { ... }`",
+        "`impl Type for Behavior`",
+        "`async fn`",
+        "Read Static And Dynamic Text Correctly",
+        "`StaticString` is the type of string literals",
+        "Passing a `StaticString` copies that\npointer-and-length view",
+        "never silently becomes `String<A>`",
+        "Learn It In One Pass",
+        "This is the shortest useful version of the language",
+        "Static text",
+        "Dynamic text preview",
+        "Counted loop",
+        "Nested loop exit",
+        "UFC loop control",
+        "Sync effect preview",
+        "Async effect preview",
+        "Sync allocator preview",
+        "Async allocator preview",
+        "Read the outer type first",
+        "`Task<Result<...>>` means the checked value belongs to scheduled work",
+        "`String<A>` is dynamic owned storage\nand must keep allocator ownership visible",
+        "Read Effects And Allocators Correctly",
+        "`Sync` and `Async` are effect modes in type surfaces",
+        "not function\nkeywords",
+        "The outer result type tells you whether work is complete",
+        "`Allocator<T, Sync>` and\n`Allocator<T, Async>` are different capabilities",
+        "Use This Mental Model",
+        "Stable Vs Preview Surface",
+        "What Is Stable Right Now",
+        "Stable today",
+        "Preview surface",
+        "The split matters most for strings",
+        "static bytes plus length",
+        "known location and\nbyte count after compilation",
+        "A dynamic string owns\nruntime memory, has capacity, can grow",
+        "Control is explicit",
+        "Text ownership is explicit",
+        "literal never silently allocates dynamic text",
+        "Effects are explicit",
+        "Behavior relationships are explicit",
+        "Tooling truth comes from the compiler",
+        "Sync/Async/Allocator Quick Rules",
+        "allocator-backed owners keep the allocator",
+        "async work returns a task-shaped value",
+        "loop handles are compiler-owned",
+        "StaticString is not a String",
+        "Static text and dynamic text are different types",
+        "loop control is prefix-only",
+        "## The Five Rules To Remember",
+        "A block result is its final expression",
+        "A string literal is `StaticString`",
+        "static bytes plus length baked into the\n   program image",
+        "Runtime-owned text is `String<A>`",
+        "Loops have one stable shape",
+        "Sync and async are visible in types",
+        "Mutable inferred local",
+        "statement-level early exits",
+        "values come from final expressions, not `return`",
+        "## Assignment And Mutation",
+        "## Operators And Casts",
+        "## Calls And UFC",
+        "value.method(args)",
+        "method(value, args)",
+        "uniform function call\nsyntax",
+        "call-site\nspellings, not alternate declaration forms",
+        "## Attached Methods",
+        "Prefer direct `Type.method = ...` declarations",
+        "## Blocks Produce Their Final Expression",
+        "Zen does not use a `return` keyword",
+        "Pattern arms are blocks",
+        "## Static And Dynamic Strings",
+        "## Error Handling",
+        "Result<T, E>",
+        "Option<T>",
+        "No exceptions",
+        "No null",
+        "## Memory And Ownership",
+        "Allocation is explicit",
+        "does not hide\nheap allocation",
+        "OwnedBytes<T, A>",
+        "pointer, length, and allocator capability",
+        "StaticString",
+        "baked into the program",
+        "compile-time-sized static text",
+        "stable pointer-and-length view",
+        "allocator-backed String",
+        "program image",
+        "does not allocate",
+        "fixed byte count known from the compiled program",
+        "copies the view, not a heap allocation",
+        "Runtime text construction belongs on an allocator-aware path",
+        "allocator-managed capacity, length, and storage",
+        "identity_static",
+        "identity_dynamic",
+        "The call site stays honest about timing",
+        "make_buffer<T, A: Allocator<T, Sync>>",
+        "source-level `String` annotations\nare gated",
+        "non-owning view",
+        "only literal bytes are\nguaranteed to be baked",
+        "does not allocate a dynamic `String`",
+        "## Loops",
+        "prefix-only: call `loop`",
+        "Loop syntax is prefix-only",
+        "`break` and `continue` are not Zen source forms",
+        "Behavior bounds can appear on generic parameters",
+        "The allocator is part of the owner",
+        "control handle, not a user-defined object",
+        "compiler-owned loop-control verbs",
+        "arbitrary user methods",
+        "The compiler recognizes only the\ncontrol verbs",
+        "not general user-defined\nmethod dispatch",
+        "complete stable loop surface",
+        "There is no suffix/body-first loop spelling",
+        "There is no hidden loop result channel",
+        "Loop control recipes",
+        "not ordinary methods named by strings",
+        "does not implement `done` or `next`",
+        "sum_to = (limit: i32) i32",
+        "Nested loops can control the current loop or an outer loop",
+        "The UFC spelling puts the control verb first",
+        "loop((l) {\n    done(l)\n    next(l)\n})",
+        "loop",
+        "l.done()",
+        "l.next()",
+        "done(l)",
+        "next(l)",
+        "outer.done()",
+        "inner.next()",
+        "## Imports And Modules",
+        "## Defer",
+        "## Gated Preview: Sync, Async, And Allocators",
+        "Current compiler paths reject\nthese spellings with feature-gate diagnostics",
+        "### Sync And Async Preview",
+        "### Allocator Preview",
+        "### Ownership Preview",
+        "Read these preview signatures literally",
+        "Sync function",
+        "Async function",
+        "Sync allocator",
+        "Async allocator",
+        "Sync",
+        "Async",
+        "Allocator<T, Sync>",
+        "Allocator<T, Async>",
+        "cannot grow, shrink, or release memory",
+        "The allocator-backed `String<A>` type is dynamic: it owns runtime memory",
+        "Sync allocation returns `Result` directly",
+        "Async allocation returns",
+        "There is no hidden conversion between those two shapes",
+        "The outer type is the\neffect boundary",
+        "a `Sync` API returns the result it computed now",
+        "an `Async` API returns a `Task<...>`",
+        "Sync code cannot call async operations without an explicit runtime boundary",
+        "allocator and scheduler APIs should expose their effect mode",
+        "not source\nkeywords placed before a function",
+        "There is no source-level `async` keyword",
+        "There is no `async fn` spelling",
+        "There is no `impl Type for Behavior` spelling",
+        "Dynamic memory ownership is visible in the returned type",
+        "bytes plus allocator ownership and an effect mode",
+        "what element type is being allocated",
+        "which capability must later release or grow the storage",
+        "`Allocator<T, Sync>` can allocate `T` now",
+        "`Allocator<T, Async>` can allocate `T` later",
+        "`Buffer<T, A>` owns memory only because `A` is kept with the buffer",
+        "String<A>",
+        "Planned `.await()`",
+        "@builtin.raw_allocate",
+        "@builtin.raw_deallocate",
+        "@builtin.raw_reallocate",
+        "## Pointer, Slice, And Array Types",
+        "RawPtr<T>",
+        "`Ptr<T>`, `MutPtr<T>`, `Slice<T>`, and `[T; N]`",
+        "raw pointer offset",
+        "comptime type matching",
+        "actor framework",
+        "gated design",
+        "docs/V1_SPEC.md",
+        "examples/05_loops.zen",
+    ] {
+        assert!(
+            guide.contains(required),
+            "Learn guide is missing expected tour or gated-preview text: {required}"
+        );
+    }
+
+    for stale in ["## Impl Blocks", "Type.impl =", ".impl = {", "impl blocks:"] {
+        assert!(
+            !guide.contains(stale),
+            "Learn guide should avoid teaching non-behavior impl-block syntax as the public tutorial path: {stale}"
+        );
+    }
+}
